move bbox_head
